packages/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rethink-llmhub"
version = "1.0.3"
description = "LLMHub CLI — manage LLM specs and runtime configs"
readme = "README.md"
authors = [
  { name = "Rethink Paradigms", email = "info@rethink-paradigms.com" },
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["llm", "ai", "llmhub", "model-management", "config", "cli"]
requires-python = ">=3.10"
dependencies = [
    "rethink-llmhub-runtime>=1.0.3",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
]

[project.scripts]
llmhub = "llmhub_cli.cli:app"

[project.urls]
"Homepage" = "https://github.com/rethink-paradigms/llm-hub"
"Documentation" = "https://github.com/rethink-paradigms/llm-hub#readme"
"Bug Tracker" = "https://github.com/rethink-paradigms/llm-hub/issues"
"Source Code" = "https://github.com/rethink-paradigms/llm-hub"

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "llmhub_cli",
    "llmhub_cli.commands",
    "llmhub_cli.catalog",
    "llmhub_cli.catalog.sources",
    "llmhub_cli.generator",
    "llmhub_cli.generator.spec",
    "llmhub_cli.generator.needs",
    "llmhub_cli.generator.selection",
    "llmhub_cli.generator.catalog_view",
    "llmhub_cli.generator.emitter",
    "llmhub_cli.tools",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
